requirements.txt
PyYAML~=6.0.2
openai~=1.51.2